backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
numpy==1.24.3
numpy-stl==3.0.1
Pillow==10.1.0
python-multipart==0.0.6

